split wiki module functions into individual files
