+++ modules/user/__init__.py
@@ -2,7 +2,7 @@ import re
 
 from core.builtins.message import MessageSession
 from core.component import on_command
-from modules.wiki.dbutils import WikiTargetInfo
+from modules.wiki.utils.dbutils import WikiTargetInfo
 from .user import get_user_info
 
 usr = on_command('user', alias=['u'],
