+++ modules/wiki/utils/ab.py
@@ -1,7 +1,7 @@
 from core.dirty_check import check
 from core.elements import Url
 from modules.wiki.utils.UTC8 import UTC8
-from modules.wiki.wikilib import WikiLib
+from modules.wiki.utils.wikilib import WikiLib
 
 
 async def ab(wiki_url):
