+++ modules/wiki/utils/ab_qq.py
@@ -2,7 +2,7 @@ from config import Config
 from core.dirty_check import check
 from core.logger import Logger
 from modules.wiki.utils.UTC8 import UTC8
-from modules.wiki.wikilib import WikiLib
+from modules.wiki.utils.wikilib import WikiLib
 
 
 async def ab_qq(wiki_url):
