+++ modules/wiki/utils/rc_qq.py
@@ -5,7 +5,7 @@ from core.dirty_check import check
 from core.logger import Logger
 from modules.wiki.utils.UTC8 import UTC8
 from modules.wiki.utils.action_cn import action
-from modules.wiki.wikilib import WikiLib
+from modules.wiki.utils.wikilib import WikiLib
 
 
 async def rc_qq(wiki_url):
