+++ modules/wiki/utils/screenshot_image.py
